+++ cli/CHANGELOG.md
@@ -21,6 +21,10 @@ _Released 3/21/2024_
 
 - Updated jose from `4.11.2` to `4.15.5`. Addressed in [#29086](https://github.com/cypress-io/cypress/pull/29086).
 
+**Bugfixes:**
+
+- Fixed an issue When there are special characters in our test folder path, sometimes the folder path is not parsed correctly.Fixes [#25839](https://github.com/cypress-io/cypress/issues/25839).
+
 ## 13.7.0
 
 _Released 3/13/2024_